VW-935 Remove total price, add production assignment, fix production schedule status

+++ Services.Manafacturing.Service/ProductionAssignment/IProductionAssignmentService.cs
@@ -10,7 +10,7 @@ namespace VErp.Services.Manafacturing.Service.ProductionAssignment
 {
     public interface IProductionAssignmentService
     {
-        Task<IList<ProductionAssignmentModel>> GetProductionAssignments(int scheduleTurnId);
+        Task<IList<ProductionAssignmentModel>> GetProductionAssignments(long scheduleTurnId);
         Task<IList<ProductionAssignmentModel>> UpdateProductionAssignment(ProductionAssignmentModel data);
         Task<IList<ProductionAssignmentModel>> CreateProductionAssignment(ProductionAssignmentModel data);
         Task<bool> DeleteProductionAssignment(int productionAssignmentId);
